Changed data transformation paradigm:  * We now create all transformations (log2, log10, none) all the time  * The transformations are _all_ normalized. That means that you can call getNoramlized(Transformation.LOG_2, ..) and getNoramlized(Transformation.NONE, ..) from the same table  * This implies that we can switch transformations at run-time, even show different transformation on a per-axis or per-view label
This still needs to be implemented in views, etc. There is still a defaultTransformation.

Fixed parallel coordinates - changed texture stuff to pixelGL..
Re-wrote gates.
Removed AGate, NominalGate -> should be handled all the same way.

Some PCs bugs remain. 
 

git-svn-id: svn://localhost:8080/caleydo_main/trunk@5769 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.core/src/org/caleydo/core/data/collection/column/container/FloatContainer.java
@@ -134,7 +134,7 @@ public class FloatContainer implements INumericalContainer<Float> {
 		for (int index = 0; index < container.length; index++) {
 			tmp = container[index];
 
-			target[index] = (float) Math.log(tmp) / (float) Math.log(base);
+			target[index] = (float) (Math.log(tmp) / Math.log(base));
 
 			if (target[index] == Float.NEGATIVE_INFINITY) {
 				target[index] = 0;
